switch to tsx
