+++ package.json
@@ -52,7 +52,7 @@
   "scripts": {
     "build": "gatsby build",
     "dev": "gatsby develop -p 3000",
-    "format": "prettier --write \"{content, src}/**/*.{js,jsx,json,md}\"",
+    "format": "prettier --write \"{content, src}/**/*.{js,jsx,json,md,tsx}\"",
     "start": "npm run dev",
     "serve": "gatsby serve",
     "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing \""
